Refactor ChatEnhancements into modular components for better organization
